# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Contract Lens API",
    "description": "Intelligent legal contract analysis platform.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "v1 API"
    }
  ],
  "components": {
    "schemas": {
      "APIErrorDetail": {
        "type": "object",
        "properties": {
          "field": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["code", "message"]
      },
      "APIErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/APIErrorDetail" }
              },
              "requestId": { "type": "string" }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "responses": {
          "200": { "description": "User registered" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/contracts": {
      "get": {
        "summary": "List contracts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of contracts" }
        }
      },
      "post": {
        "summary": "Upload a contract",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contract uploaded and analyzed" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/APIErrorResponse" }
          }
        }
      }
    }
  }
}